requirements.txt
boto3==1.4.4
protobuf==3.3.0
google-cloud==0.25.0
